+++ image_builder/image_builder/build.py
@@ -73,9 +73,12 @@ def build_job_image(
 
         if config.clean_up_workspaces:
             with wrap_context('cleaning up the workspace'):
+                update_deployment_phase(config, deployment_id, 'cleaning up the workspace')
                 if repo_dir.exists():
                     shutil.rmtree(repo_dir)
 
+        logger.info(f'finished building an image {manifest.name}, deployment ID: {deployment_id}')
+        update_deployment_phase(config, deployment_id, 'finalizing the build')
         return image_names, logs, error
 
     finally: