+++ lifecycle/lifecycle/job/audit.py
@@ -33,8 +33,8 @@ class AuditLogger:
             properties=properties_json,
             username_executor=username_executor,
             username_subject=username_subject,
-            fatman_name=job_name,
-            fatman_version=job_version,
+            job_name=job_name,
+            job_version=job_version,
         )
         ale.save()
 
@@ -50,8 +50,8 @@ class AuditLogger:
 @db_access
 def read_audit_log_user_events(
     username: Optional[str] = None,
-    fatman_name: Optional[str] = None,
-    fatman_version: Optional[str] = None,
+    job_name: Optional[str] = None,
+    job_version: Optional[str] = None,
 ) -> List[AuditLogEventDto]:
     """
     Get list of active public endpoints that can be accessed
@@ -59,8 +59,8 @@ def read_audit_log_user_events(
     :param job_version: Exact job version or an alias ("latest" or wildcard)
     """
     username_filter = Q(username_executor=username) | Q(username_subject=username) if username else Q()
-    job_name_filter = Q(fatman_name=fatman_name) if fatman_name else Q()
-    job_version_filter = Q(fatman_version=fatman_version) if fatman_version else Q()
+    job_name_filter = Q(job_name=job_name) if job_name else Q()
+    job_version_filter = Q(job_version=job_version) if job_version else Q()
     queryset = models.AuditLogEvent.objects.filter(
         username_filter & job_name_filter & job_version_filter
     ).order_by('-timestamp')[:100]