+++ lifecycle/lifecycle/job/dto_converter.py
@@ -8,14 +8,14 @@ from racetrack_commons.entities.dto import AuditLogEventDto, PublicEndpointReque
 from racetrack_commons.entities.dto import JobDto, JobFamilyDto, DeploymentDto
 
 
-def job_family_model_to_dto(model: models.FatmanFamily) -> JobFamilyDto:
+def job_family_model_to_dto(model: models.JobFamily) -> JobFamilyDto:
     return JobFamilyDto(
         id=model.id,
         name=model.name,
     )
 
 
-def job_model_to_dto(model: models.Fatman, config: Config) -> JobDto:
+def job_model_to_dto(model: models.Job, config: Config) -> JobDto:
     return JobDto(
         id=model.id,
         name=model.name,
@@ -56,8 +56,8 @@ def esc_model_to_dto(model: models.Esc) -> EscDto:
 
 def public_endpoint_request_model_to_dto(model: models.PublicEndpointRequest) -> PublicEndpointRequestDto:
     return PublicEndpointRequestDto(
-        fatman_name=model.fatman.name,
-        fatman_version=model.fatman.version,
+        job_name=model.job.name,
+        job_version=model.job.version,
         endpoint=model.endpoint,
         active=model.active,
     )
@@ -73,6 +73,6 @@ def audit_log_event_to_dto(model: models.AuditLogEvent) -> AuditLogEventDto:
         properties=properties,
         username_executor=model.username_executor,
         username_subject=model.username_subject,
-        fatman_name=model.fatman_name,
-        fatman_version=model.fatman_version,
+        job_name=model.job_name,
+        job_version=model.job_version,
     )